requirements.txt
aniso8601==10.0.1
bcrypt==4.0.1
beautifulsoup4==4.14.2
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
curl_cffi==0.13.0
dnspython==2.7.0
email-validator==2.3.0
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security==5.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
frozendict==2.4.6
google-generativeai==0.8.5
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multitasking==0.0.12
numpy==2.0.2
pandas==2.3.3
passlib==1.7.4
peewee==3.18.2
platformdirs==4.4.0
protobuf>=3.20.2,<6.0.0dev
pycparser==2.23
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
requests==2.32.5
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
WTForms==3.2.1
yfinance==0.2.66
zipp==3.23.0
